+++ doc.go
@@ -1,8 +1,10 @@
 /*
 Package haversine provides the great circle distance between two points
-on the surface of the earth.  Points are identified by latitude and
-longitude, and distance results are returned in nautical miles.  Functions
-are also provided to return the distance in kilometers or statute miles.
+on a sphere. Several functions return great circle distances of the Earth.
+Points are identified by latitude and longitude in degrees, and distance
+results are returned in nautical miles.  Functions are also provided to
+return the distance in kilometers or statute miles, or the geocentric
+angular separation between those points.
 
 The below example shows how to calculate the shortest path between two
 coordinates on the surface of the Earth.
@@ -11,6 +13,7 @@ coordinates on the surface of the Earth.
 
     import (
         "fmt"
+	"math"
 
         "github.com/FATHOM5/haversine"
     )
@@ -22,12 +25,14 @@ coordinates on the surface of the Earth.
         }
         paloAlto := haversine.Coord{ // Palo Alto, California
             Lat: 37.4419,
-            Lon: 122.1430,
+            Lon: -122.1430,
         }
 
-        nm := haversine.Distance(austin, paloAlto)
+        nm := haversine.DistanceNM(austin, paloAlto)
 
         fmt.Printf("%.1f miles is a long walk to Silicon Valley.\n", nm)
+        fmt.Printf("it's only %.1f miles on the moon, though.\n", haversine.Distance(austin, paloAlto, 937.9))
+	fmt.Printf("That's an angle of %.1f degrees\n", 180/math.Pi*haversine.IntAngle(austin, paloAlto))
     }
 */
 package haversine